[FIX] tvl logic
