+++ projects/neutra-finance/index.js
@@ -1,19 +1,22 @@
-const ADDRESSES = require('../helper/coreAssets.json')
-const { sumTokens2 } = require('../helper/unwrapLPs')
+async function tvl(_, _b, _cb, { api }) {
+  const nGlpVaults = "0x6Bfa4F1DfAfeb9c37E4E8d436E1d0C5973E47e25";
+  const nUSDCVault = "0x2a958665bC9A1680135241133569C7014230Cb21";
 
-async function tvl(_, _b, _cb, { api, }) {
-  const vaults = ['0x2a958665bc9a1680135241133569c7014230cb21']
-  const tokens = await api.multiCall({ abi: 'address:token', calls: vaults })
-  const bals = await api.multiCall({ abi: 'uint256:totalAssets', calls: vaults })
-  api.addTokens(tokens, bals)
-  return sumTokens2({
-    api,
-    owners: ['0x6bfa4f1dfafeb9c37e4e8d436e1d0c5973e47e25'],
-    tokens: [ADDRESSES.arbitrum.fsGLP, ADDRESSES.optimism.DAI,],
-  })
+  const nGlpTotalValue = await api.call({
+    abi: "uint256:totalValue",
+    target: nGlpVaults,
+  });
+  const nUSDCTotalValue = await api.call({
+    abi: "uint256:totalAssets",
+    target: nUSDCVault,
+  });
+
+  const tvl = nUSDCTotalValue + (nGlpTotalValue / 1e24).toFixed();
+
+  return tvl; // tvl in 6 decimal
 }
 module.exports = {
   arbitrum: {
     tvl,
-  }
+  },
 };
